Bluewing update for improved Unix compatibility
Added explicit casts to make it obvious we're discarding data, where the compiler warned about it.
Added some debug logging to lw_eventqueue. Event queue seems to work well under Blue Android Client, but not Blue Unix Server, although Android uses Unix code...
Added some error checking to ::pipe(), as recommended. May add recommendations for write() and read(), but they're buried deep in void-returning functions, so they can wait.

Switched some initialization into C++ style. I plan to later roll liblacewing code back to mixed C/C++ style, so external code can invoke via C calls, but that requires some workarounds for std::string_view and std::atomic, so won't be happening quite yet.

+++ DarkEdif/Lib/Shared/Lacewing/src/global.c
@@ -120,7 +120,7 @@ void lw_dump (const char * buffer, size_t size)
 	size_t i;
 	int row_offset = 0, row_offset_c = 0, row = 0;
 
-	if (size == -1)
+	if (size == (size_t)-1)
 		size = (lw_ui32) strlen (buffer);
 
 	fprintf (stderr, "=== " lwp_fmt_size " bytes @ %p ===\n", size, buffer);
